requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
coverage==7.3.2
dj-database-url==2.1.0
dj-rest-auth==2.1.9
Django==3.2.23
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.4
django-resized==1.0.2
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
sqlparse==0.4.4
uritemplate==4.1.1
whitenoise==6.6.0
